CustomFileSystemModel: "grandparent sibling" checking
There were problems with grandparents.  Given the directory layout:

    A-
      -B-
         -C-

if A was already checked, and the user checked C, A would remain checked
(instead of becoming partially checked, which is the desired behaviour of
tarsnap-gui).

If we didn't check every ancestor, B could end up unchecked, whereas we want
it to be partially checked (as well as A being partially checked).