elidedlabel.cpp, filepickerdialog.cpp: clang-format
The difference in the constructor initializer line comes from

    ConstructorInitializerAllOnOneLineOrOnePerLine: true

so that is clearly deliberate.

The change to

    connect() ... [&]() {

is probably not desired, since `BreakBeforeBraces: Allman` is supposed to
always break before braces.  My guess is that clang-format is confused by the
anonymous function.  However, I see that `[&]() {` is used in many other
places in tarsnap-gui, so I guess this is an acceptable format.  (And I
absolutely approve of taking the output of automatic formatting tools, whether
or not it's the "ideal" format.)