CustomFileSystemModel: recurse into subdirectories
This is a tricky one, and requires a recursive solution.  For readability,
here's the scenario:

    ## Grandfather directory: select grandson, then select grandfather.
    actions:
    + dir-1/dir-2/file-5.txt
    + dir-1

    results:
    2 dir-1
    1 dir-1/dir-2
    1 dir-1/dir-2/file-5.txt
    1 dir-1/dir-2/file-6.txt
    1 dir-1/file-3.txt
    1 dir-1/file-4.txt
    0 file-1.txt
    0 file-2.txt

When we select dir-1, we *must* recurse downwards through the tree so that we
can unselect file-5.txt.  To avoid travelling through directories
unnecessarily, we only recurse into directories which were previously marked
as Checked or PartiallyChecked.