Use download/upload icons, and arrow right
The "pancakes" icon doesn't fit with the style for other icons, so
it would be best not to use that for backup.  The best replacement
I could find was the "upload" icon.  And if we switch to "upload",
then we might as well use "download" for the restore operation.

I thought we'd try arrow-circle-right for creating a job, although
I admit that there's less justification for that particular
choice.