Display copy&pasteable commands
The previous commands would fail if a directory contained a space
(for example, with the default ".../Tarsnap Backup Inc./...").
QProcess automatically escapes spaces in arguments before it runs
a command, but unbelievably there is no way to access the
"escaped" arguments.  (QProcess.arguments() returns the un-escaped
strings)

Adding single quotes around each argument seems like a safer
solution than trying to escape spaces.

NOTE: the actual command that QProcess runs is *not* changed by
this commit.  This commit only changes what is displayed on the
command-line and console log.