JobWidget: make infoLabel red upon error
I didn't notice the text for about 10 seconds; I kept on clicking
a box in the file tree, then un-clicking it, and wondering why the
file tree moved up and down.

I think more of a re-think of this widget's GUI is in order, but
having a more visible error message is still a good place to
start.