Improve startup sequence for automatic sched. backup.
Check for Tarsnap servers reachability before starting a backup
and if not reachable allow for an extra 30 seconds before trying
again and then finally quitting with an explicit message (log and
system notification) of why the backup has not proceeded.

Trying to backup as soon as the system wakes up is not graceful
enough to account for wifi autojoin or other cases of system load,
thus many backup failures.