tests: exit if there's a failure
"Oh, I won't bother trying to implement my own test suite", said Graham.  "Qt
has a whole QTest infrastructure.  I'm sure that will be powerful and easy to
use."

sigh.

And yes, I keep on doubting whether this is the right way to go, but googling
suggests that the alternative to running a whole bunch of independent tests (as
we're doing now) is to compile them all into one large binary that runs each
test class independently.  I'm not necessarily opposed to that, but tying
things together with shell scripts seems a safer way to keep things
independent.