Proper fix for "about" window (including test)
Trying to have two completely different ways of showing the "About" window was
making this messy.  Let's use the same basic method (i.e. clicking on the
"About Tarsnap" PushButton in the Help tab) on all OSes, and add an extra way
of getting the "About" window for OSX.

We might be able to eliminiate aboutMenuClicked() by using some fancier
connect() logic.  My primary goal here is clarity and having a working test; I
may revisit mainwindow.cpp in the future.  The test should only change if we
change our minds about the desired behaviour.

Even though aboutMenuClicked() is only useful for OSX, we cannot add
    #ifndef Q_OS_OSX
to mainwindow.h because that confuses the moc, which apparently doesn't handle
built-in defines until 5.8.0:
    https://bugreports.qt.io/browse/QTBUG-34593