TarsnapTask: always generate a finished() signal
Otherwise, the GUI won't be notified that a task has finished with a failure
(for example, trying to register with the server if tarsnap-keygen fails).

An alternate method for resolving this problem would be to connect the
canceled() and/or dequeue() signals to everything that tries to run a task, but
that would significantly complicate TaskManager, so I opted for this instead.