Use optparse to parse command-line arguments
QCommandLineParser requires a QCoreApplication or QApplication to exist.
That doesn't make it useful if we want to look at the command-line arguments
and decide if we need a GUI or not.  :(

We want to be able to have a QCoreApplication (command-line only) with --check,
because otherwise we won't be able to run --check as part of the brew install
script.  Fortunately, we have libcperciva's magic getopt.  (An added bonus of
using C for this is that it compiles and runs quicker than C++.)

The only downside of using C for parsing command-line arguments is that we lose
Qt's tr() translation mechanism.  But I think that's an acceptable trade-off,
especially since few people will manually using the command-line arguments
anyway.