CoreApplication, AppCmdline: _appDataDir -> _configDir
This variable (and --appdata command-line option) was added in
9c380161443d43594b299c33b2d6a4dd1b702b59.

As far as I can tell, it's only used in

    settings.setPath(QSettings::IniFormat, QSettings::UserScope, _appDataDir);

so I think this value is actually the config dir, not the app data dir.

The --appdata argument has NOT been updated yet.  When I add a replacement
argument, I'll mark --appdata as deprecated (to be removed in the next major
version of tarsnap-gui).