init-shared: add init_shared_settings()
Right now it's not necessary, but in the future we'll want to be able to return
a message from this function.

I considered adjusting struct init_info to return a QStringList of messages,
but then I realized that I'd want a different status to go along with some
messages (i.e. "is this a warning or error message?"), so then I started
thinking about a linked list... and at that point I decided it was getting more
complicated than it was worth.