ConfirmationDialog: validate confirmation text
Ideally, we wouldn't show the "Ok" button unless the full confirmation text was
given.  Unfortunately, that would require either manually creating the QDialog
ourselves (instead of using QInputDialog), or else messing with low-level Qt
stuff like QWidget::layout() in order to access the buttons.

I'm not ruling out manually creating the confirmation text dialog ourselves,
but it's not a priority, and changing the text on the button gets us 80% of the
way there with only 20% of the effort.