Translations: downgrade to v1.0
I've learned a better way of updating translations, so I'm going back and
fixing this.

This commit performs
	lupdate Tarsnap.pro
on d5afbc2ef4cc612b7b22d5549e00214f86cf187b (which is v1.0, 2017-07-27).

Everything in this commit is undoing good changes that we want to end up with,
but in the interests of keeping track of what's happening, I think it's better
to go back to a stable start position (i.e. v1.0) and move forward from there.
All of these changes will be re-made by the next set of commits.