tests: add valgrind framework
Qt (and the C++ libraries it uses) leaks memory like a sieve, so it's
useful to generate a suppression file using code that we *know* should
be ok.  This is a technique which is used in other Tarsnap software.

- minimal:       tests stuff that will be used in qtest-minimal.
- qtest-minimal: the smallest example which involves QTest.

This does not yet enable valgrind on any tests; that will be done slowly
and carefully in the future.