ConsoleLog: singleton -> a global variable
The previous code used a singleton.  Many blog and forum posts have been
written about whether singletons are a good design pattern or not, and I
don't plan to add to that discussion.

I was considering removing it because explicit initialization and
clean-up is easier for me understand (because I'm not an expert C++
programmer), and also easier for compilers and code analysis tools.  But
that wasn't an important enough reason to change working code.

The deciding factor behind this switch was thinking about how
Tarsnap-GUI fits into the Tarsnap company code.  Our other projects use
"C with implicit classes" (via functions with opaque pointers), so I
think the GUI fits in better if it was basic C++, i.e. "C with classes".

After all, the "cost" of having a global variable instead of a singleton
is a 1-line initialization function, and a 1-line clean-up function.