CustomFileSystemModel: rename "index" variable to "idx"
Recent versions of Qt marked QModelIndex::child() as deprecated; we
should use index() instead.  However, our subclass used index as a
variable, following the suggestion of the Qt docs.  (Why on earth would
the official API docs use the same name for variables and a function?!?)

This resolves the naming clash, which allows the "avoid using the
deprecated function" commit to be much more clear.