TaskManager, SetupDialog: add sanity check for existing keyfile
The existing code relied on the key either existing or not, but I'd like
to use the existing check of which option the user selected.  It's
likely that this check will never be needed, but if I mess up something
while refactoring the SetupDialog, I'd rather get an error at this point
in the code.