task-defs: move some #defines here
This begins consolidating Tarsnap-related tasks into src/tasks/.  These are
generally the command-line operations, but a few other tasks will also be
included (such as checking that the client has an Internet connection).

The hard rule is that nothing in src/tasks/ should have side effects (which
includes using connect() or declaring a class).

Future note: I'm not 100% certain that the CMD_* definitions are useful, but
I'll keep them for now so that I'm not making too many changes at once.