task-utils: new copies of TaskManager utils
These functions are not used yet, but will be in the next commit.  I'm
adding them separately so that it's easier to see the refactoring
involved in that commit.

- makeTarsnapCommand() is almost identical to the one in taskmanager.cpp.
- makeTarsnapArgs() is very similar to initTarsnapArgs() in
  taskmanager.cpp, but differs in that it doesn't take `QStringList &args`
  as a parameter.

I'm leaving the related functions in taskmanager.cpp untouched.  As code
moves into the task/ subdirectory, I'll rewrite it to use the task-utils
functions instead.  Once the whole process is finished, the related
functions in taskmanager.cpp will be deleted.