AppGui: use pointer for _journal
Previously, quitting AppGui triggered the destructor for Journal, which
attempts to write
    log("==Session end==");
which triggeres a
    PersistentStore &store = getStore();

In most cases, this is appropriate, but not when the user attempts to
quit the SetupDialog without setting anything up.

Using a pointer means that we can check that _journal has been
initialized before we delete it.