Re-run initializeCore() upon completing the setup
If we cancelled the setup wizard, quit as per normal.

If we ran it normally or skipped it, we need to restart the
initialization for reasons that I don't fully understand.

FUTURE: I'm not fond of AppGui handling both the main app and the setup.
I'm considering adding an AppSetup, then having main.cpp decide which
variant to run -- AppCmdline, AppGui, or AppSetup.  Of course, we'd need
to have the latter two inside a loop, so that the setup can proceed
gracefully into the main app, which can then proceed back into the setup
wizard if requested.