tests: replace "-platform offscreen" with QT_QPA_PLATFORM=offscreen
The QGuiApplication (inherited by QApplication) modifies argc and argv
to remove the "-platform offscreen" pair.  That makes it tricky to pass
that option to multiple QApplications (in sequence, not parallel) in the
same binary.

Thankfully, the docs note:
    When in doubt, use the environment variables.
    https://doc.qt.io/qt-5/qguiapplication.html#QGuiApplication
so there should be no harm in switching to QT_QPA_PLATFORM.