ElidedLabel: add messageNormal(), messageError()
The main goal is to abstract away the UI-specific handling, so that the
"business logic" layer doesn't need to deal with it.

FUTURE 1: it's likely that the precise visual representation (i.e. colour)
will change.

FUTURE 2: at the moment, this is changing the object's whole stylesheet.
Another option would be to use Qt's dynamic properties, as per
    https://stackoverflow.com/q/1595476
    https://wiki.qt.io/DynamicPropertiesAndStylesheets

However, as noted by those links, a change in the "dynamic" property
will not take effect unless you force it to reload the style.  The
example in the above wiki is:
    myLineEdit->setProperty("urgent", true);
    myLineEdit->style()->unpolish(myLineEdit);
    myLineEdit->style()->polish(myLineEdit);
which looks very ugly.

At the moment, I'm hesitant to switch to such a half-heartedly supported
option.  I'm open to investigating this again in the future, especially
when Qt 6 is out.