TWizard, TWizardPage: add replacements for QWizard, QWizardPage
QWizard is almost exactly what I wanted for SetupWizard (the replacement
for SetupDialog), but a number of things are too hard-coded.  For
example:
- it's impossible to increase the font size of the title, and the logo
  isn't displayed unless there's a subtitle.
- there's some corner cases in keyboard navigation, wherein a navigation
  button can retain visual focus while a widget inside a QWizardPage is
  being edited.
After many hours of investigating workarounds and existing Qt bug
reports (many of which were deemed "out of scope"), I decided that it
would be easier to write my own classes.

TWizard and TWizardPage keep everything easily modifiable.  They were
written specifically for my use-case, so they diverge from QWizard and
QWizardPage in a few important aspects:

- pages must be added all at once via TWizard::addPages(), instead of
  one-by-one with QWizard::addPage().
- it does not support the field() and registerField() mechanism.
- it only supports a linear wizard, rather than non-linear ones.
- it does not support subtitles.
- it has 4 buttons; from left to right: skipButton, backButton,
  nextButton, and finishButton.  Connections are created "automatically"
  via TWizardPage::connect_ui_buttons().