SetupWizard: new classes based on TWizard
I think I've reached the point of diminishing returns for keeping this
as a separate branch.  I'm sure that there will be further tweaks and
minor fixes required, but it's time to merge it.

The most important thing is that (as you can see in the next few
commits) we can replace SetupDialog with SetupWizard in the tests and
the main application by only changing those nouns, plus a few minor
tweaks to #define macros in the tests.

Known fixes needed:
- rename some elements on the CliPage:
  - "Advanced" to "Details".
  - remove prepended "cli" from UI objects.
- rework the RegisterPage UI:
  - move the machine name outside of the "choice" widget.
  - use a QTabWidget (with CSS to put the tabs in the center) instead
    of a QStackedWidget + two QPushButtons.
- clean up the RegisterPage code:
  - the above UI changes should allow it to be much simpler.

I held off on the above changes because they would involve changing the
test suite (thus foiling the "replace the implementation with minimal
changes to the interface"), or because it might be useful to have that
history in the new setupwizard files (rather than the soon-to-be-deleted
SetupDialog files).