Use SetupWizard
This replaces the old SetupDialog with the new SetupWizard, but we keep
the old class for now, to simplify updates to the translations.

Note that the adjustments to .cpp files are replacing
    SetupDialog -> SetupWizard
and editing two #define constructs in the tests.

FUTURE: I'm keeping SetupDialog in the tree for a bit longer, and plan
to remove it once I've finished refactoring SetupWizard.