tests/valgrind: omit all leaks using libfontconfig
Either that library, or Qt's use of that library, is a leaky disaster.
In the old system, I tried to keep track of how each leak originated,
but I've given up now.  For my purpose of "did I make a mistake writing
new code?", I think that it's sufficient to ignore all
libfontconfig-related leaks.