tests/valgrind: add qtest-gui-weird
These appear to resolve the SetupWizard leak, but I admit that I'm not
certain what's happening.  I also expect that these are non-minimal
examples.  However, I've spent more time than I wanted to on this, so
I'm adding it "as-is" and moving on.

FUTURE: I plan to use this file as a dumping ground for messy valgrind
stuff, and move stuff to other directories if/when I think that I've
reduced it to a minimal example.