OkLabel: add rich text
I want to add an "ok" display to a centered QLabel (actually,
ElidedLabel, but same problem occurs).  Unfortunately, there's some
spacing weirdness[1], so it's easier to append some rich text to the
existing label, instead of having a separate OkLabel.

[1] maybe the width of a QLabel is calculated including all characters
involved in rich text?  That was the issue which prompted
ElidedAnnotatedLabel.