Utils: remove old validate functions
SetupDialog is no longer used[1], so all that really matters is that it
compiles.

[1] I'm keeping it for a bit longer in case I discover a problem in the
new Setup Wizard and want to look at the old code.  (Yes, in that case I
could play games with git history, but I figured that it would be
simpler if I just postponed deleting the file for another month or two.)