TWizardPage: auto-focus "proceed" button on initialization
Despite the <tabstops> section, it appears that the "first focus" widget
varies based on Qt version.  This avoids that uncertainty.

It may seem odd to focus on a button without checking that it's enabled
(e.g., what if the "proceed" button was disabled by checkComplete()?),
but this appears to be fine.