StopTasksDialog: reset cancellation handling
This should have been part of c238b81640e2fa267309f6af7d3d3fc4f9f3c341.

I thought that fixing this part of the code would be easy, so I did that
at the same time as moving the code from MainWindow to StopTasksDialog.
I was wrong.

QMessageBox::finished() is not documented as diverging from
QDialog::finished(), but it does.  Instead of returning
QDialog::Accepted or QDialog::Rejected, it returns a
QDialogButtonBox::StandardButton code.  This is not impossible to work
around, but it deserves to have its own commit, rather than that change
being buried in the middle of refactoring.