StopTasksDialog: create a new cancel button every time
This should have been part of 747943bde56962243c95dea47769b9f77a735e5d.

Fixes the broken "make test_visual" in tests/mainwindow.

There was an unfortunate combination of factors:
- QMessageBox does not reset clickedButton() when you start a new exec()
  or open().
- QMessageBox does not provide any way of resetting clickedButton() manually.
- the test suite uses QDialog::accept() and QDialog::reject().
- when a test uses QMetaObject::invokeMethod(..., Qt::QueuedConnect),
  it doesn't take effect when at a VISUAL_WAIT (as I'd expected).
  That meant that this bug did not manifest with a regular "make test".

To be fair, QMessageBox was probably not intended to be a window that we
kept on re-using, and especially not re-using with the QDialog-level
accept() and reject().