StopTasksDialog: clean up old cancel button
This could have been part of 3d70f9f7422513d794f1465e399a7026c9be23e5,
but it's not terrible to have it as a separate commit for emphasis.

Fixes the broken "make test_valgrind" in tests/small-widgets.

When we remove the button, we also need to free the memory.  However, if
we do that before the next addButton(), then the next cancel button
might be allocated in exactly the same memory location as the old one,
meaning that if we've accept()'d the StopTasksDialog, clickedButton()
can erroneously return the memory location of the current cancel button.
(And as mentioned in that commit, there's no way to clear clickedButton().)