RegisterPage: hard-code QTabWidget height
The goal is to avoid having the QTabWidget changing the height based on
the tab.  This shouldn't be too hard, and online discussions suggest
that an unchanging height is the default behaviour (which they want to
change).  I suspect that I'm missing something about how QLayouts work.

I'm not happy with hard-coding the height, but it works, and I can
always revisit it later.