StopTasksDialog: don't use Qt::QueuedConnection
I don't understand why this was originally queued.  This arose from
    connect(&actionMenu, &QMenu::triggered, &_stopTasksDialog, &QDialog::accept,
            Qt::QueuedConnection);
in src/widgets/mainwindow.cpp, when this functionality was there.  The
Qt::QueuedConnection was originally added in:
    c0628068aedb1248a35c1feaa9fae6fec6993ac1
and the whole thing was moved into StopTasksDialog in:
    c238b81640e2fa267309f6af7d3d3fc4f9f3c341

All of the tests still pass, and it looks ok with manual testing.  I'm
still a bit hesitant to remove something that I don't understand, but
it's time to bite the bullet and move on.