tests: make VISUAL_INIT() also ->show() on non-visual tests
This will allow us to simplify the MainWindow::quit_tasks() tests, as
well as generally increasing the test coverage.  I didn't enable this
initially because that added test coverage meant more valgrind problems,
but it appears to be ok now.

This commit also removes the specially-enabled offscreen show().

I considered renaming VISUAL_INIT() to INIT() or INIT_SHOW(), but I
decided that it wasn't worth all the extra lines of diff.