Doxygen: don't try to name member groups
It appears that in modern doxygen, a member group can either:
  - apply a doc string to each thing inside it, thereby avoiding the "not
    documented" warning,
OR
  - have a title for the grouping, but don't things inside it as being
    documented due to that.

It's possible that a later version of doxygen will go back to the
previos behaviour ("group title is sufficient to be considered
'documented'), and/or I can achieve that behaviour by tweaking the
Doxygen file.  For now, it's more important to me that doxygen not
produce a whole bunch of false warnings on this repository.