PersistentObject: make it a Q_OBJECT
The design of "PersistentObject non-Qt, but Archive, Job, & Journal all
Qt objects" was present in its very first appearance:
    2015-04-08 8c4513ad91d195db8658099430869fcf59cd88a6
and of course still there when it was moved into src/:
    2015-05-26 0be7be0ddac7918919325f9ad1eea27065d0e816

I've failed to think of a reason why it should be this way.  I'm wary of
Chesterton's fence, but I'm going to go ahead and change it.  If
necesary, I can come back and revert this commit later.