JobDetailsWidget: prep for "early bail" formatting
Don't run `make format` on this commit.  I'm not a fan of the diff that
git produces in ::canSaveNew(), so I've isolated the `make format` to
completely its own commit.

The problem is that git is preserving the indentation on
             _ui->infoLabel->setStyleSheet("#infoLabel { color: darkred; }");
and
             _ui->infoLabel->show();
(along with some lines containing "{ } else"), which makes it very
un-obvious that the "format" commit is merely removing one level of
indentation.