SetupDialog: remove
This hasn't been used since 64a300bc2d29a27946f6327618a48b66192da4ad.

I kept it around for a while in case I had to check anything, but I
think it's safe to delete now.  And it's in the git history, of course,
so I can still check it if need be.