Prepare for const QString &, wrt QString.remove()
This awkwardness arises because QString.remove() modifies itself, in
addition to returning a pointer to its modified form.  As a result, we
can't call .remove() on a const QString &str.

One solution would be to leave the function alone (and trust that Qt is
doing the right thing when it comes to strings and copying memory --
which, to be fair, doesn't sound like a terrible idea), but my
preference is to make the API as clean as possible.  If that means a bit
of extra ugliness inside two .cpp files, so be it.