ArchiveDetailsWidget: use fixed size for metadataWidget
It should be possible to achieve this with QLayout or SizePolicy or
something like that, but so far I've been unable to do it.  I'd love to
revisit this if/when I discover how to say "make this widget's height as
high as it needs to be, but no higher" without manually specifying a
number of pixels.  :(