CmdlineTask: remove stderr from monitoring
I introduced a subtle bug in:
    2021-04-11 CmdlineTask: add setMonitorOutput()
    ac0c1ed4709c9c095618dce809268d9b4b0f9327

Namely, if you're monitoring the QProcess::readyReadStandardError signal
(or the ...Output signal), the message from that signal can be in a
queue while you destroy the underlying object.  If that happens, we face
a race condition: can the slot process the message before the object is
deleted?

At a very rough guess, we lost this race around 30% of the time when
running TestTaskManager::fail_registerMachine_empty_key() on FreeBSD 12.2.
So far, I have not observed the test failing on Linux, either on my own
hardware or via GitHub Actions; but that could just be a coincidence.

The point of that PR was to monitor the --fsck output so that we could
show a progress bar during the setup wizard.  That is not changed by
this commit; we're still monitoring stdout.

As far as I know, there's no problem if we miss a little bit of stdout
output, so I'm not planning on changing the stdout monitoring.  However,
I will continue to work on the test suite and may need to revisit this.