MainWindow: always show the menu bar
On macOS, there should be no change -- the menu bar will still be at the
top, as the system menu bar.

On window managers which do not have a system menu bar, this will add a
menu bar to the top of the window.

On window managers which have a shared system menu bar... I'm not
certain what this commit will do.  The QMenuBar documentation claims
that Ubuntu Unity will have a shared menu bar.  Later on, it states
that:
     if the com.canonical.AppMenu.Registrar service is available on the
     D-Bus session bus, then Qt will communicate with it to install the
     application's menus into the global menu bar
     https://doc.qt.io/qt-5/qmenubar.html

However, I have not observed tarsnap-gui having a global menu bar on
either Ubuntu 20.04 or Ubuntu 16.04.  It's possible that D-Bus was
broken on one or both of those virtualbox images -- I haven't been
careful about keeping those images up-to-date, and I've ignored all
warning messages on those systems.

I've decided to live with that uncertainty and move on.  As long as the
menu bar appears in one place, I'll leave it up to Qt & the window
manager to decide which one.