test: skip (some) offscreen tests on macOS with Qt < 5.15
There is a segfault in Qt on macOS in QTabBar (which is used by
QTabWidget).  It was fixed here:
    https://github.com/qt/qtbase/commit/3197932e6fb03fb9c0ff669af858cb94e3836d79
but that commit is marked as "Pick-to: 5.15".

One example of that breakage is in tests/app-setup when we did:
    2020-02-22 Setupwizard Register: switch to TabWidget
    c923b54f45371f49d06fb35509953dca1172ec8f
(The same problem exists in JobsTabWidget and MainWindow.)

This commit does not attempt to release allocated memory if we bail or
skip; I fully expect that valgrind would report leaks in such
situations.  To the best of my knowledge, the tests should still be fine
on macOS with Qt 5.15 or later.