MainWindow: always create 3 macOS Window QActions
This reduces the amount of OS-specific code, meaning that compiling on
any particular platform (in my case, FreeBSD) will cover more
functionality.

On the minus side, this means that non-macOS binaries will be slightly
larger and use slightly more memroy, as they now contain unnecessary
functionality (allocating and initializing those three actions).  On
FreeBSD 12.2 with Qt 5.15.2, the stripped binary is 2384 bytes larger
after this commit, which is insignificant.