MainWindow: take ownership of ConsoleLogDialog
This was previously part of HelpWidget.

Also, add a menu item to show the console log.

Regarding the tests: I opted to keep them in the HelpWidget test for
now, instead of moving them to the MainWindow test.  This is primarily
because we can't run the MainWindow tests on macOS, as discussed in
    09f3ea2932a59b4a2631cac5746c093b3fc064d8