lib/widgets: add TTabWidget
The current implementation requires that TTabWidget is given a "large"
and "small" image, and that those images won't change for the lifetime
of the object.

It wouldn't be hard to relax either constraint, but this was all that we
need for Tarsnap-GUI, so I refrained from spending more time on it.