MainWindow: add a File menu with a Quit action (non-macOS)
This does not apply to macOS, because on that platform the
"Quit Tarsnap" action is automatically put into the "app menu".

I'm deeply uncertain about calling the menu "File" rather than "Tarsnap"
or some kind of "app menu".  In the end, my reasoning was based on
discoverability:
- anybody who's used a non-macOS computer in the past couple of decades
  will expect to find a "Quit" or "Exit" open inside a "File" menu.
- people who are are familiar with macOS or some modern applications
  might expect to find "Quit" in the left-most menu (whatever that menu
  is called), but that's not everybody.

People familiar with modern GUI design might think that "File->Quit" is
a bit archaic or un-stylistish, but they won't find it confusing.