style: add parentheses for "return"
In libcperciva C style, we use parentheses for "return".  Let's reduce
the amount of unnecessary differences in our code bases.

For future reference, check this with:
    git grep "  return " -- ':(exclude)*.py' | grep -v "return ("

(I'm not fond of the formatting on lines 56-57 of src/dir-utils.cpp, but that's
a trivial quibble and not worth fighting with clang-format at this time.)