TTabWidget: have persistent QPixmap
The previous version of this class assumed that we would only set the
logos once.  That was enforced by a pair of Q_ASSERTs, but I remove
those in
    2023-02-24 TTabWidget: fix width recalculation logic
    8cd81725c170e23f8edbd61be2cacd40a3c52e8e

When TTabWidget is used in an application, it's not unreasonable to
disallow changing the logos; but when we're developing with it in the
Qt designer, that's not ideal.