TPath*Browse: don't use system spacing
The macOS style has default spacing of 8 pixels, whereas fusion (the
default style for linux) has a default spacing of 6 pixels.

At the present time, we don't require the whole GUI to look
pixel-identical on every platform.  However, we _would_ like the
fixed-size "library widgets" to have the same size.