MainWindow: macOS: don't remove the "empty" File menu
This fixes a bug introduced in:
    2022-04-25 MainWindow: add a File menu with a Quit action (non-macOS)
    e5e3cd81bb2d079439e4c26e10b637cacc7b4846

QMenuBar on macOS has a "menu bar merging feature" which puts actionQuit
into the "Application Menu" (the left-most menu, not counting the
"apple" menu).

I assumed that meant that actionQuit would be removed from
_ui->menu_File, so we would be left with an empty File menu
(undesirable).

However, it turns out that although Qt [1] moves actionQuit
*graphically*, in terms of object ownership, actionQuit is still inside
menu_File.  As a result, the ->setvisible(false) was disabling the Quit
action.

[1] This is with Qt 5.15.10 on macOS 13.6.1.