tests-mocked/fake-cmdline: add
This fake `tarsnap` and `tarsnap-keygen` could be called "mocks",
"fakes", or "stubs" (wikipedia notes that these terms are used
inconsistently in the literature).

The important thing is that these replace the normal `tarsnap`
command-line interface, allowing us to fake certain operations without
involving the Tarsnap server.  This allows us to test aspects of the GUI
much faster and more comfortably.